requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
requests==2.31.0
tiktoken==0.5.2
python-dotenv==1.0.0
weaviate-client==4.5.5
sentence-transformers==2.7.0
langchain==0.1.16
langchain-community==0.0.34
litellm==1.35.0
redis==5.0.1
celery==5.3.4
transformers==4.40.0
scikit-learn==1.4.2
pandas==2.2.1
numpy==1.26.4
imbalanced-learn==0.12.0
beautifulsoup4==4.12.3
lxml==5.1.0
slack-sdk==3.27.1
requests-oauthlib==1.4.0
msal==1.26.0
zeep==4.2.1
pymongo==4.6.1
sqlalchemy==2.0.28
psycopg2-binary==2.9.9
pymysql==1.1.0
cryptography==42.0.5
bcrypt==4.1.2
PyJWT==2.8.0
